% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregation}
\alias{aggregation}
\title{Fonction qui permet d'aggréger un jeu de données avec une colonne date, en un jeu de donnée avec un pas de temps plus grand}
\usage{
aggregation(dd, form)
}
\arguments{
\item{dd}{dataset a aggréger}

\item{form}{format pour l'aggrégation qui doit être parmis : "Jeu de données", "Heure", "Jour", "Mois"}
}
\value{
le dataset aggrégé
}
\description{
Fonction qui permet d'aggréger un jeu de données avec une colonne date, en un jeu de donnée avec un pas de temps plus grand
}
