% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{aspi_noaa}
\alias{aspi_noaa}
\title{Téléchargement des données à partir du serveur ftp du NOAA, format ISD lite}
\usage{
aspi_noaa(dd, station)
}
\arguments{
\item{dd}{jeu de données avec à minima une colonne date}

\item{station}{station sélectionnée à partir de l'interface shiny}
}
\value{
un jeu de données avec les dates en première colonne et les températures correspondantes
}
\description{
Téléchargement des données à partir du serveur ftp du NOAA, format ISD lite
}
