% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_uncertainty.R
\name{cusum}
\alias{cusum}
\title{Calcul la carte de controle cusum pour analyser les dérives lentes}
\usage{
cusum(dd, date_deb_ref, date_fin_ref, incertit)
}
\arguments{
\item{dd}{jeu de données avec à minima les colonnes date, energie et prediction}

\item{date_deb_ref}{date de début de la période de référence}

\item{date_fin_ref}{date de fin de la période de référence}

\item{incertit}{erreur type : incertitude du modèle divisée par la racine du nombre de valeurs}
}
\value{
un dataframe avec les dates, limites de la carte de contrôle et les valeurs SI+ et SI- de la CUSUM
}
\description{
Calcul la carte de controle cusum pour analyser les dérives lentes
}
