% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Jeu de données de consommation d'un bâtiment sur l'île de la Réunion}
\format{
105 000 lignes avec 5 colonnes:
\describe{
\item{Col1}{Date}
\item{Col2}{Energie}
\item{Col3}{Température}
\item{Col4}{DJU}
\item{Col5}{Horaires d'ouverture}
}
}
\usage{
dataset
}
\description{
Relevé de consommation entre deux périodes au pas de temps 15 minutes
}
\keyword{datasets}
