% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ.R
\name{economie}
\alias{economie}
\title{Calcul l'économie estimée en fin d'année}
\usage{
economie(dd, date_deb_ref, date_fin_ref)
}
\arguments{
\item{dd}{jeu de données avec a minima une colonne date, une colonne énergie et une colonne prédiction}

\item{date_deb_ref}{date de début de la période de référence}

\item{date_fin_ref}{date de fin de la période de référence}
}
\value{
un dataframe avec la date et le pourcentage d'économie
}
\description{
Calcul l'économie estimée en fin d'année
}
