% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusum_uncertainty.R
\name{incertitude}
\alias{incertitude}
\title{Calcul l'incertitude type du modèle}
\usage{
incertitude(dd, date_deb_ref, date_fin_ref, econ)
}
\arguments{
\item{dd}{jeu de données avec a minima une variable, une colonne energie, une colonne prédiction et une variable indépendante}

\item{date_deb_ref}{date de début de la période de référence}

\item{date_fin_ref}{date de fin de la période de référence}

\item{econ}{économie attendue en pourcentage}
}
\value{
l'incertitude selon la méthode hat matrix écrite dans la FDX30-148
}
\description{
Calcul l'incertitude type du modèle
}
