% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launcher.R
\name{launch_mpeb}
\alias{launch_mpeb}
\title{Lance l'interface de l'outil MPEB}
\usage{
launch_mpeb()
}
\value{
l'interface shiny lancée dans un navigateur web
}
\description{
Lance l'interface de l'outil MPEB
}
