% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_country.R
\docType{data}
\name{list_country}
\alias{list_country}
\title{Liste des pays pour le téléchargement des données météos sur le site du NOAA}
\format{
Data frame avec la correspondance des identifiants de pays pour les données météorologiques :
\describe{
\item{Col1}{FIPS.ID}
\item{Col2}{COUNTRY.NAME}
}
}
\usage{
list_country
}
\description{
Relevé de consommation entre deux périodes au pas de temps 15 minutes
}
\keyword{datasets}
