% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_list.R
\docType{data}
\name{liste_station}
\alias{liste_station}
\title{Liste des stations météorologiques datant de décembre 2019}
\format{
30 000 lignes avec 11 colonnes:
\describe{
\item{Col1}{USAF}
\item{Col2}{WBAN}
\item{Col3}{STATION.NAME}
\item{Col4}{CTRY}
\item{Col5}{STATE}
\item{Col6}{ICAO}
\item{Col7}{LAT}
\item{Col8}{LON}
\item{Col9}{ELEV.M.}
\item{Col10}{BEGIN}
\item{Col11}{END}
}
}
\usage{
liste_station
}
\description{
Liste des stations météorologiques datant de décembre 2019
}
\keyword{datasets}
