% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{modelisation}
\alias{modelisation}
\title{Modélise la consommation énergétique}
\usage{
modelisation(dd, model_type, date_deb_ref, date_fin_ref, econ)
}
\arguments{
\item{dd}{jeu de données avec au moins une colonne date et une colonne énergie}

\item{model_type}{modèle statistique à appliquer au choix entre "Regression simple", "Regression par morceaux" et "TOWT"}

\item{date_deb_ref}{date de début de la période de référence}

\item{date_fin_ref}{date de fin de la période de référence}

\item{econ}{économie attendue avec la réalisation de l'AAPE}
}
\value{
une liste comprenant le jeu de données d'entrée auquel on a rajouté une colonne prédiction et un objet correspondant à la régression qui a été réalisée
}
\description{
Modélise la consommation énergétique
}
