% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occup.R
\name{occupation}
\alias{occupation}
\title{Fonction qui retourne le dataset avec une colonne supplémentaire d'occupation/inoccupation}
\usage{
occupation(dd)
}
\arguments{
\item{dd}{jeu de données avec au moins une colonne date, une variable température et une colonne énergie}
}
\value{
le jeu de données d'entrée avec en plus une colonne occup qui vaut 0 ou 1 en fonction de l'occupation prédite par le modèle
}
\description{
Fonction qui retourne le dataset avec une colonne supplémentaire d'occupation/inoccupation
}
