% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{select_ss_station}
\alias{select_ss_station}
\title{Filtrage des stations par le pays sélectionné dans l'interfaces}
\usage{
select_ss_station(stations, pays)
}
\arguments{
\item{stations}{liste des stations}

\item{pays}{pays qui provient de l'interface shiny}
}
\value{
stations météorologiques actives dans le pays sélectionné lors de la plage de date du jeu de données
}
\description{
Filtrage des stations par le pays sélectionné dans l'interfaces
}
