% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{select_station}
\alias{select_station}
\title{Sélectionne les stations en activité durant la période du jeu de données}
\usage{
select_station(stations, date_deb, date_fin)
}
\arguments{
\item{stations}{liste des stations}

\item{date_deb}{première date du jeu de données}

\item{date_fin}{dernière date du jeu de données}
}
\value{
un dataframe avec les stations qui étaient en activité durant la plage de temps du jeu de données
}
\description{
Sélectionne les stations en activité durant la période du jeu de données
}
