% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistiques.R
\name{statistiques}
\alias{statistiques}
\title{Renvoie les statistiques du modèle}
\usage{
statistiques(dd, date_deb_ref, date_fin_ref, reg, var, econ)
}
\arguments{
\item{dd}{jeu de données}

\item{date_deb_ref}{date de début de la période de référence}

\item{date_fin_ref}{date de fin de la période de référence}

\item{reg}{objet régression de R qui provient de la fonction model}

\item{var}{variable qui vaut 0 si le calcul du modèle n'a pas été fait, 1 sinon}

\item{econ}{économie attendue à la suite d'une AAPE}
}
\value{
Un tableau avec les paramètre RMSE, CVRMSE, r², r² ajusté et p-value du modèle
}
\description{
Renvoie les statistiques du modèle
}
